[IR] Account for dispatch receiver type when selecting fake overrides
Out of multiple candidate fake overrides we selected the most specific
one based on their return types. However, if the return types matched,
we just selected the first candidate, which was not good, since it'd
make the semantics depend on the order of base classes/interfaces
in the inheritance clause.

Instead, let's first look at the return types, and then, if they are
the same, at the dispatch receiver types.

If we still couldn't find a single candidate, give up and throw an
exception.

#KT-52553 Fixed